backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# QR Code generation
qrcode[pil]==7.4.2
Pillow==10.1.0

# Multi-cloud storage
boto3==1.34.0
azure-storage-blob==12.19.0
azure-identity==1.15.0

# HTTP client for external requests
httpx==0.25.2
aiofiles==23.2.1

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Monitoring and metrics
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# Logging and configuration
structlog==23.2.0
python-json-logger==2.0.7
python-dotenv==1.0.0

# Database (optional for metadata storage)
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0

# Security and authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Health checks and utilities
psutil==5.9.6
